+++ projects/ngx-form-object/src/lib/form-object-builder/form-object-builder.ts
@@ -1,28 +1,29 @@
-import { FormBuilder, ValidatorFn, AbstractControl } from '@angular/forms';
+import { AbstractControl, FormBuilder, ValidatorFn } from '@angular/forms';
 import { ExtendedFormArray } from '../extended-form-array/extended-form-array';
 import { ExtendedFormControl } from '../extended-form-control/extended-form-control';
 import { FormObject } from '../form-object/form-object';
 import { FormStore } from '../form-store/form-store';
 import { capitalize } from '../helpers/helpers';
 import { PropertyOptions } from '../interfaces/property-options.interface';
-import { MODEL_BUILD_CONTROL_METHODS } from '../types/model-metadata.type';
-import { CREATE_FORM_OBJECT_METHODS } from '../types/model-metadata.type';
-import { Injectable } from '@angular/core';
+import { CREATE_FORM_OBJECT_METHODS, MODEL_BUILD_CONTROL_METHODS } from '../types/model-metadata.type';
 
-@Injectable()
-export class FormObjectBuilder {
-	public formBuilder: FormBuilder = new FormBuilder();
+export class FormObjectBuilder<T> {
+	public formBuilder: FormBuilder;
 
-	public create(formObject: FormObject): FormStore {
-		const formFields = {};
+	constructor() {
+		this.formBuilder = new FormBuilder();
+	}
+
+	public create(formObject: FormObject<T>): FormStore<T> {
+		const formFields: Record<string, AbstractControl> = {};
 
 		Object.assign(formFields, this.createAttributeFormFields(formObject));
 		Object.assign(formFields, this.createHasManyFormFields(formObject));
 		Object.assign(formFields, this.createBelongsToFormFields(formObject));
 
-		const formStoreClass: any = formObject.formStoreClass ? formObject.formStoreClass : FormStore;
+		const formStoreClass = formObject.formStoreClass ? formObject.formStoreClass : FormStore;
 
-		const formStore: FormStore = new formStoreClass(
+		const formStore: FormStore<T> = new formStoreClass(
 			formFields,
 			formObject.formGroupOptions.validator,
 			formObject.formGroupOptions.asyncValidator
@@ -32,11 +33,11 @@ export class FormObjectBuilder {
 		return formStore;
 	}
 
-	private createAttributeFormFields(formObject: FormObject): Record<string, unknown> {
-		const attributeFormFields = {};
+	private createAttributeFormFields(formObject: FormObject<T>): object {
+		const attributeFormFields: Record<string, AbstractControl> = {};
 
-		formObject.attributePropertiesKeys.forEach((attributeName: string | symbol) => {
-			const buildFunction = this.getBuildFunction(formObject, attributeName);
+		formObject.attributePropertiesKeys.forEach((attributeName: string) => {
+			const buildFunction = formObject[`build${capitalize(attributeName.toString())}`];
 			const validators: ValidatorFn | Array<ValidatorFn> = formObject.getValidators(attributeName.toString());
 			const maskFunction: Function = formObject[`mask${capitalize(attributeName.toString())}`];
 
@@ -52,10 +53,10 @@ export class FormObjectBuilder {
 		return attributeFormFields;
 	}
 
-	private createHasManyFormFields(formObject: FormObject): Record<string, unknown> {
-		const hasManyFormFields = {};
+	private createHasManyFormFields(formObject: FormObject<T>): object {
+		const hasManyFormFields: Record<string, AbstractControl> = {};
 
-		formObject.hasManyPropertiesKeys.forEach((propertyName) => {
+		formObject.hasManyPropertiesKeys.forEach((propertyName: string) => {
 			const buildFunction = this.getBuildFunction(formObject, propertyName);
 			const validators: ValidatorFn | Array<ValidatorFn> = formObject.getValidators(propertyName.toString());
 			const hasManyModels = formObject.model[propertyName];
@@ -69,7 +70,7 @@ export class FormObjectBuilder {
 		return hasManyFormFields;
 	}
 
-	private createBelongsToFormFields(formObject: FormObject): Record<string, unknown> {
+	private createBelongsToFormFields(formObject: FormObject<T>): object {
 		const belongsToFormFields = {};
 
 		formObject.belongsToPropertiesKeys.forEach((propertyName: string | symbol) => {
@@ -103,7 +104,7 @@ export class FormObjectBuilder {
 		return belongsToFormFields;
 	}
 
-	private getBuildFunction(formObject: FormObject, propertyName: string | symbol): () => AbstractControl {
+	private getBuildFunction(formObject: FormObject<T>, propertyName: string | symbol): () => AbstractControl {
 		const propertyNameString = propertyName.toString();
 
 		// Deprecated in favour of build control decorator
@@ -117,7 +118,7 @@ export class FormObjectBuilder {
 	}
 
 	private buildRelationshipModels(
-		formObject: FormObject,
+		formObject: FormObject<T>,
 		relationshipName: string | symbol,
 		relationshipModels: Array<any> = [],
 		propertyOptions: PropertyOptions
@@ -126,7 +127,11 @@ export class FormObjectBuilder {
 		const formGroups: Array<any> = [];
 
 		relationshipModels.forEach((relationshipModel) => {
-			const formStore: FormStore = this.createRelationshipFormObject(formObject, relationshipName, relationshipModel);
+			const formStore: FormStore<T> = this.createRelationshipFormObject(
+				formObject,
+				relationshipName,
+				relationshipModel
+			);
 			if (formStore) {
 				formGroups.push(formStore);
 			}
@@ -143,11 +148,11 @@ export class FormObjectBuilder {
 	}
 
 	private createRelationshipFormObject(
-		formObject: FormObject,
+		formObject: FormObject<T>,
 		relationshipName: string | symbol,
 		relationshipModel: any,
 		propertyOptions: PropertyOptions = {}
-	): FormStore {
+	): FormStore<T> {
 		const relationshipNameString: string = relationshipName.toString();
 
 		// Deprecated in favour of create form object decorators
@@ -158,13 +163,13 @@ export class FormObjectBuilder {
 		}
 
 		if (createFormObjectFunction) {
-			const modelFormObject: FormObject = createFormObjectFunction.call(
+			const modelFormObject: FormObject<T> = createFormObjectFunction.call(
 				formObject,
 				relationshipModel,
 				null,
 				propertyOptions
 			);
-			const formStore: FormStore = this.create(modelFormObject);
+			const formStore: FormStore<T> = this.create(modelFormObject);
 			return formStore;
 		} else {
 			// There is no function specified for creating form object for ${relationshipName.toString()}